Use Database class, rename script to analyze
