+++ pyproject.toml
@@ -26,7 +26,7 @@ pre-commit = "^3.2.2"
 distance = "waste.distance:main"
 ingest = "waste.ingest:main"
 simulate = "waste.simulate:main"
-analyze = "waste.analyze_results:main"
+analyze = "waste.analyze:main"
 
 [tool.black]
 line-length = 79
